Promote ZiplineManifest to the main zipline module
I'd like to either add ZiplineManifest as a parameter to
EventListener, or add it as an optional member on Zipline.
In either case promoting it to zipline is working towards
that goal.

Related: https://github.com/cashapp/redwood/issues/640
